feature: UserDao인터페이스 추가 및 기존 클래스 네이밍 변경
