+++ src/main/java/dao/CountingDaoFactory.java
@@ -9,8 +9,8 @@ import javax.sql.DataSource;
 @Configuration
 public class CountingDaoFactory {
     @Bean
-    public UserDao userDao() {
-        return new UserDao(dataSource());
+    public UserDaoJdbc userDao() {
+        return new UserDaoJdbc(dataSource());
     }
 
     @Bean
