+++ src/main/java/dao/DaoFactory.java
@@ -21,8 +21,8 @@ public class DaoFactory {
     }
 
     @Bean
-    public UserDao userDao() {
-        return new UserDao(dataSource());
+    public UserDaoJdbc userDao() {
+        return new UserDaoJdbc(dataSource());
     }
 
     @Bean
