+++ src/main/java/dao/UserDao.java
@@ -1,52 +1,13 @@
 package dao;
 
 import domain.User;
-import org.springframework.jdbc.core.JdbcTemplate;
-import org.springframework.jdbc.core.RowMapper;
 
-import javax.sql.DataSource;
-import java.sql.ResultSet;
-import java.sql.SQLException;
 import java.util.List;
 
-public class UserDao {
-    private JdbcTemplate jdbcTemplate;
-
-    public UserDao(DataSource dataSource) {
-        this.jdbcTemplate = new JdbcTemplate(dataSource);
-    }
-
-    public void add(User user) throws SQLException {
-        jdbcTemplate.update("insert into users(id, name, password) values (?,?,?)",
-                user.getId(), user.getName(), user.getPassword());
-    }
-
-    public void deleteAll() throws SQLException {
-        jdbcTemplate.update("delete from users");
-    }
-
-    public User get(String id) throws SQLException {
-        return jdbcTemplate.queryForObject("select * from users where id =?", userMapper(), id);
-    }
-
-    public List<User> getAll() {
-        return jdbcTemplate.query("select * from users order by id", userMapper());
-    }
-
-    public int getCount() throws SQLException {
-        return jdbcTemplate.queryForObject("select count(*) from users", Integer.class);
-    }
-
-    private RowMapper<User> userMapper() {
-        return new RowMapper<User>() {
-            @Override
-            public User mapRow(ResultSet rs, int rowNum) throws SQLException {
-                User user = new User();
-                user.setId(rs.getString("id"));
-                user.setName(rs.getString("name"));
-                user.setPassword(rs.getString("password"));
-                return user;
-            }
-        };
-    }
+public interface UserDao {
+    void add(User user);
+    User get(String id);
+    List<User> getAll();
+    void deleteAll();
+    int getCount();
 }
