+++ src/main/java/dao/UserDaoJdbc.java
@@ -1,0 +1,52 @@
+package dao;
+
+import domain.User;
+import org.springframework.jdbc.core.JdbcTemplate;
+import org.springframework.jdbc.core.RowMapper;
+
+import javax.sql.DataSource;
+import java.sql.ResultSet;
+import java.sql.SQLException;
+import java.util.List;
+
+public class UserDaoJdbc implements UserDao{
+    private JdbcTemplate jdbcTemplate;
+
+    public UserDaoJdbc(DataSource dataSource) {
+        this.jdbcTemplate = new JdbcTemplate(dataSource);
+    }
+
+    public void add(User user) {
+        jdbcTemplate.update("insert into users(id, name, password) values (?,?,?)",
+                user.getId(), user.getName(), user.getPassword());
+    }
+
+    public void deleteAll() {
+        jdbcTemplate.update("delete from users");
+    }
+
+    public User get(String id) {
+        return jdbcTemplate.queryForObject("select * from users where id =?", userMapper(), id);
+    }
+
+    public List<User> getAll() {
+        return jdbcTemplate.query("select * from users order by id", userMapper());
+    }
+
+    public int getCount() {
+        return jdbcTemplate.queryForObject("select count(*) from users", Integer.class);
+    }
+
+    private RowMapper<User> userMapper() {
+        return new RowMapper<User>() {
+            @Override
+            public User mapRow(ResultSet rs, int rowNum) throws SQLException {
+                User user = new User();
+                user.setId(rs.getString("id"));
+                user.setName(rs.getString("name"));
+                user.setPassword(rs.getString("password"));
+                return user;
+            }
+        };
+    }
+}
